requirements.txt
# Core Data Engineering
apache-airflow==3.0.3rc5
apache-airflow-providers-docker==3.13.0
dbt-core==1.9.0
dbt-postgres==1.9.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31

# API & Web
requests==2.32.3

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.3.1
pytest-cov==5.0.0
pytest-mock==3.14.0

# Code Quality
black==24.4.2
flake8==7.1.0
isort==5.13.2
pylint==3.2.3

# Type Checking
mypy==1.10.1
types-requests==2.32.0.20240622

# Documentation
sphinx==7.3.7
sphinx-rtd-theme==2.0.0
